Allow the Plugin to be used when using `-host-header=rewrite`
Currently using `ngrok http -host-header=rewrite dev.site:80` does not work with this plugin
as the `$_SERVER['HTTP_HOST']` is set to `http://dev.site` so will not be picked upt by current checks.

this commit Updates the if statement to also check `$_SERVER['HTTP_X_ORIGINAL_HOST']` for the `NGROK_URL`
and makes the necessary chnages to deliver the correct base_url and therefore the correct assets.

the reason this is required is for multistore sites that set the mage_run_code based on the  URL.